classes utilisateur et responsable
