+++ App/ProtoControllers/Responsable/ATraitement.php
@@ -95,88 +95,6 @@ abstract class ATraitement
     }
 
     /**
-     * Retourne les informations d'un utilisateur
-     *
-     * @param string $login
-     * 
-     * @return string $donnees
-     */
-    public static function getDonneesUtilisateur($login)
-    {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT *
-                FROM conges_users 
-                WHERE u_login = \''.  \includes\SQL::quote($login).'\'';
-        $query = $sql->query($req);
-        $donnees = $query->fetch_array();
-
-        return $donnees;
-    }
-
-    /**
-     * Retourne l'id des groupes d'un responsable
-     *
-     * @param string $resp
-     * 
-     * @return array $ids
-     */
-    public function getIdGroupeResp($resp)
-    {
-        $ids = [];
-
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT gr_gid AS id FROM `conges_groupe_resp` WHERE gr_login =\''.$resp.'\'';
-        $res = $sql->query($req);
-
-        while ($data = $res->fetch_array()) {
-            $ids[] = (int) $data['id'];
-        }
-
-        return $ids;
-    }
-
-    /**
-     * Retourne l'id des groupes d'un grand responsable
-     *
-     * @param string $gresp
-     * 
-     * @return array $ids
-     */
-    public function getIdGroupeGrandResponsable($gresp)
-    {
-        $ids=[];
-         $sql = \includes\SQL::singleton();
-         $req = 'SELECT ggr_gid AS id FROM `conges_groupe_grd_resp` WHERE ggr_login =\''.$gresp.'\'';
-         $res = $sql->query($req);
-
-         while ($data = $res->fetch_array()) {
-             $ids[] = (int) $data['id'];
-         }
-
-         return $ids;
-    }
-
-    /**
-     * Retourne le login des membres d'une liste de groupes
-     *
-     * @param array $groupesId
-     * 
-     * @return array $users
-     */
-    public function getUsersGroupe(array $groupesId)
-    {
-         $sql = \includes\SQL::singleton();
-         $req = 'SELECT gu_login FROM `conges_groupe_users` WHERE gu_gid IN (' . implode(',', $groupesId) . ')';
-         $res = $sql->query($req);
-
-         while ($data = $res->fetch_array()) {
-             $users[] = $data['gu_login'];
-         }
-
-         return $users;
-    }
-        
-    /**
      * Traitement d'une validation avec modification du solde
      * 
      * @param int $demandeId
@@ -235,7 +153,7 @@ abstract class ATraitement
             $fin    = date('H\:i', $demande['fin']);
             $duree  = \App\Helpers\Formatter::Timestamp2Duree($demande['duree']);
             $id = $demande['id_heure'];
-            $infoUtilisateur = $this->getDonneesUtilisateur($demande['login']);
+            $infoUtilisateur = \App\ProtoControllers\utilisateur::getDonneesUtilisateur($demande['login']);
             $solde = \App\Helpers\Formatter::Timestamp2Duree($infoUtilisateur['u_heure_solde']);
             $Table .= '<tr class="'.($i?'i':'p').'">';
             $Table .= '<td><b>'.$infoUtilisateur['u_nom'].'</b><br>'.$infoUtilisateur['u_prenom'].'</td><td>'.$jour.'</td><td>'.$debut.'</td><td>'.$fin.'</td><td>'.$duree.'</td><td>'.$solde.'</td>';
@@ -271,61 +189,6 @@ abstract class ATraitement
     }
 
     /**
-     * Vérifie si un utilisateur est bien le responsable d'un employé
-     * 
-     * @param string $resp
-     * @param string $user
-     * 
-     * @return bool
-     */
-    public function isRespDeUtilisateur($resp, $user) {
-        return $this->isRespDirect($resp, $user) || $this->isRespGroupe($resp, $this->getGroupesId($user));
-    }
-    
-    /**
-     * Vérifie si un utilisateur est bien le grand responsable d'un employé
-     * 
-     * @param string $resp
-     * @param array $groupesId
-     * 
-     * @return bool
-     */
-    public function isGrandRespDeUtilisateur($resp, array $groupesId) {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT EXISTS (
-                    SELECT ggr_gid
-                    FROM conges_groupe_grd_resp
-                    WHERE ggr_gid IN (\'' . implode(',', $groupesId) . '\')
-                        AND ggr_login = "'.\includes\SQL::quote($resp).'"
-                )';
-        $query = $sql->query($req);
-
-        return 0 < (int) $query->fetch_array()[0];
-    }
-
-    /**
-     * Verifie si un utilisateur est responsable d'une liste de groupe
-     * 
-     * @param string $resp
-     * @param array $groupesId
-     * 
-     * @return bool
-     */
-    public function isRespGroupe($resp, array $groupesId)
-    {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT EXISTS (
-                    SELECT gr_gid
-                    FROM conges_groupe_resp
-                    WHERE gr_gid IN (\'' . implode(',', $groupesId) . '\')
-                        AND gr_login = "'.\includes\SQL::quote($resp).'"
-                )';
-        $query = $sql->query($req);
-
-        return 0 < (int) $query->fetch_array()[0];
-    }
-    
-    /**
      * Verifie si la demande n'a pas déja été traité
      * 
      * @param string $statutDb
@@ -338,94 +201,6 @@ abstract class ATraitement
         return ($statut != \App\Models\AHeure::STATUT_ANNUL || $statut != \App\Models\AHeure::STATUT_OK || $statut != \App\Models\AHeure::STATUT_REFUS);
     }
 
-    /**
-     * Verifie si un utilisateur est grand responsable d'une liste de groupe
-     *
-     * @param string $gResp
-     * @param int $groupesId
-     * 
-     * @return bool
-     */
-    public function isGrandRespGroupe($gResp, $groupesId)
-    {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT EXISTS (
-                    SELECT ggr_gid
-                    FROM conges_groupe_grd_resp
-                    WHERE ggr_gid IN (\'' . implode(',', $groupesId) . '\')
-                        AND ggr_login = "'.\includes\SQL::quote($gResp).'"
-                )';
-        $query = $sql->query($req);
-
-        return 0 < (int) $query->fetch_array()[0];
-    }
-
-    /**
-     *  Verifie si un utilisateur est responsable d'un employé
-     * 
-     * @param string $resp
-     * @param string $user
-     * 
-     * @return bool
-     */
-    public function isRespDirect($resp, $user)
-    {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT EXISTS (
-                    SELECT u_resp_login 
-                    FROM conges_users 
-                    WHERE u_login ="'.\includes\SQL::quote($user).'"
-                        AND u_resp_login ="'.\includes\SQL::quote($resp).'"
-           )';
-    $query = $sql->query($req);
-
-    return 0 < (int) $query->fetch_array()[0];
-    }
-
-    /**
-     * retourne les identifiants de groupe auquel un utilisateur appartient
-     * 
-     * @param string $user
-     * 
-     * @return array $ids
-     */
-    public function getGroupesId($user)
-    {
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT gu_gid AS id
-                    FROM conges_groupe_users 
-                    WHERE gu_login ="'.\includes\SQL::quote($user).'"';
-        $res = $sql->query($req);
-        while ($data = $res->fetch_array()) {
-            $ids[] = (int) $data['id'];
-        }
-
-        return $ids;
-    }
-
-    /**
-     * Vérifie si le groupe d'un employé est en double validation
-     * 
-     * @param string $user
-     * 
-     * @return bool
-     */
-    protected function isDoubleValGroupe($user)
-    {
-        $groupes = [];
-        $groupes = $this->getGroupesId($user);
-        $sql = \includes\SQL::singleton();
-        $req = 'SELECT EXISTS (
-                    SELECT g_double_valid
-                    FROM conges_groupe
-                    WHERE g_gid ='. $groupes[0] . '
-                    AND g_double_valid = "Y"
-                )';
-        $query = $sql->query($req);
-
-        return 0 < (int) $query->fetch_array()[0];
-    }
-    
     /**
      * Retourne le nombre de demande en cours d'un responsable
      *
